refactor(logic): split incremental_candidate_validator into part-of concern files (Refs #3290)
Phase 0 decomposition of incremental_candidate_validator.dart (648 physical
lines, over the ≤500-line target). Prefix-replay validators and economy-
projection helpers move into two part-of concern files sharing the library's
private scope via extensions, following the deal_matcher Phase 0 pattern.

No public API, behaviour, or determinism change — pure structural split.
