+++ packages/colonizethis_logic/lib/src/orders/incremental_candidate_validator.dart
@@ -1,15 +1,3 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-
-import '../constants.dart';
-import '../diplomacy/diplomacy_resolver.dart';
-import '../economy/economy_riches_to_treasury.dart';
-import '../world/player_view.dart';
-import 'order_resolution_context.dart';
-import 'order_validators.dart';
-import 'unit_type_helpers.dart';
-import 'validator_bundle.dart';
-
 /// Incremental candidate validation primitive used by the order suggestion API
 /// to evaluate single candidates against an already-accepted `basePrefix`
 /// without running full-pass `validatePlayerOrdersWithContext`.
@@ -23,6 +11,30 @@ import 'validator_bundle.dart';
 /// full-pass validator over `basePrefix ⊕ c` and inspecting `c`'s result.
 /// This file covers all candidate types used by suggestion probes, including
 /// stateful build/work/diplomatic validators via accepted-prefix replay.
+///
+/// Prefix-replay validators and economy-projection helpers live in `part of`
+/// concern files (`incremental_candidate_validator_prefix_replay.dart`,
+/// `incremental_candidate_validator_projection.dart`) to keep each file below
+/// the repo file-size policy (`SPEC/program/dart-file-non-comment-line-size.md`);
+/// they share this library's private scope so behaviour is unchanged (Refs
+/// #3290 Phase 0 file decomposition).
+library;
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import '../diplomacy/diplomacy_resolver.dart';
+import '../economy/economy_riches_to_treasury.dart';
+import '../world/player_view.dart';
+import 'order_resolution_context.dart';
+import 'order_validators.dart';
+import 'unit_type_helpers.dart';
+import 'validator_bundle.dart';
+
+part 'incremental_candidate_validator_prefix_replay.dart';
+part 'incremental_candidate_validator_projection.dart';
+
 class IncrementalCandidateValidator {
   IncrementalCandidateValidator._({
     required this.game,
@@ -293,356 +305,4 @@ class IncrementalCandidateValidator {
         .validateNavalMission(candidate, previousRejected: false)
         .isAccepted;
   }
-
-  /// Validates a [RecruitWorkerOrder] candidate against accepted recruit
-  /// worker orders in [basePrefix] (Refs #2692 S7,
-  /// SPEC/program/order-suggestions.md § Recruit worker orders).
-  ///
-  /// Mirrors the order-engine recruit worker phase: existing recruit orders
-  /// in [basePrefix] are replayed in submission order against the player's
-  /// snapshot worker pool / stockpile / treasury so the candidate sees the
-  /// post-prefix peasant reservation ledger.
-  bool isRecruitWorkerAccepted(RecruitWorkerOrder candidate) {
-    final player = _player();
-    if (player == null) return false;
-    if (_cachedRecruitWorkerPrefixReplaySucceeded == false) {
-      return false;
-    }
-    if (_cachedPostRecruitWorkerPrefixEconomy == null) {
-      final prefixValidator = RecruitWorkerOrderValidator.withProjectedEconomy(
-        player: player,
-        stockpile: player.stockpile,
-        treasury: player.treasury,
-        workerPool: player.workerPool,
-      );
-      final existing =
-          basePrefix.recruitWorkerOrdersByPlayerId[playerId] ??
-          const <RecruitWorkerOrder>[];
-      for (final order in existing) {
-        final result = prefixValidator.validate(order, previousRejected: false);
-        if (!result.isAccepted) {
-          _cachedRecruitWorkerPrefixReplaySucceeded = false;
-          return false;
-        }
-      }
-      _cachedRecruitWorkerPrefixReplaySucceeded = true;
-      _cachedPostRecruitWorkerPrefixEconomy = (
-        stockpile: prefixValidator.stockpile,
-        treasury: prefixValidator.treasury,
-        workers: prefixValidator.workers,
-      );
-    }
-    final snap = _cachedPostRecruitWorkerPrefixEconomy!;
-    final candidateValidator = RecruitWorkerOrderValidator.withProjectedEconomy(
-      player: player,
-      stockpile: Stockpile(
-        quantities: Map<String, int>.from(snap.stockpile.quantities),
-      ),
-      treasury: snap.treasury,
-      workerPool: snap.workers,
-    );
-    return candidateValidator
-        .validate(candidate, previousRejected: false)
-        .isAccepted;
-  }
-
-  bool isBuildAccepted(BuildUnitOrder candidate) {
-    final player = _player();
-    if (player == null) return false;
-    if (_cachedBuildPrefixReplaySucceeded == false) {
-      return false;
-    }
-    final builds =
-        basePrefix.buildUnitOrdersByPlayerId[playerId] ??
-        const <BuildUnitOrder>[];
-    if (_cachedPostBuildPrefixEconomy == null) {
-      final prefixValidator = BuildOrderValidator.withProjectedEconomy(
-        game: game,
-        player: player,
-        stockpile: player.stockpile,
-        treasury:
-            player.treasury +
-            pendingRichesTreasuryDelta(
-              stockpile: player.stockpile,
-              richesCashMultiplier: game.richesCashMultiplier,
-            ),
-        workerPool: player.workerPool,
-      );
-      for (final existing in builds) {
-        final result = prefixValidator.validate(
-          existing,
-          previousRejected: false,
-        );
-        if (!result.isAccepted) {
-          _cachedBuildPrefixReplaySucceeded = false;
-          return false;
-        }
-      }
-      _cachedBuildPrefixReplaySucceeded = true;
-      _cachedPostBuildPrefixEconomy = (
-        stockpile: prefixValidator.stockpile,
-        treasury: prefixValidator.treasury,
-        workers: prefixValidator.workers,
-      );
-    }
-    final snap = _cachedPostBuildPrefixEconomy!;
-    final candidateStockpile = Stockpile(
-      quantities: Map<String, int>.from(snap.stockpile.quantities),
-    );
-    final candidateValidator = BuildOrderValidator.withProjectedEconomy(
-      game: game,
-      player: player,
-      stockpile: candidateStockpile,
-      treasury:
-          snap.treasury +
-          pendingRichesTreasuryDelta(
-            stockpile: candidateStockpile,
-            richesCashMultiplier: game.richesCashMultiplier,
-          ),
-      workerPool: snap.workers,
-    );
-    return candidateValidator
-        .validate(candidate, previousRejected: false)
-        .isAccepted;
-  }
-
-  bool isWorkAccepted(WorkOrder candidate) {
-    final player = _player();
-    if (player == null) return false;
-    if (_cachedWorkPrefixReplaySucceeded == false) {
-      return false;
-    }
-    final prefix = _ensurePostWorkPrefixState(player);
-    if (prefix == null) {
-      return false;
-    }
-    final workValidator = createWorkOrderValidator(
-      game: game,
-      player: player,
-      playerId: playerId,
-      resolution: (
-        view: view,
-        unitsById: unitsById,
-        provinceById: view.provincesById,
-      ),
-      topology: topology,
-      diplomaticOrders: diplomaticOrders,
-      tileMapByRegion: tileMapByRegion,
-      civilianDraftMoveUnitIds: _civilianDraftMoveUnitIds(),
-      devExclusiveTiles: Set<String>.from(prefix.devExclusive),
-      stockpile: Stockpile(
-        quantities: Map<String, int>.from(prefix.stockpile.quantities),
-      ),
-      treasury: prefix.treasury,
-      factionMembership: _factionMembership(),
-      initialSeenUnitIds: prefix.seenUnitIds,
-    );
-    return workValidator
-        .validate(candidate, previousRejected: false)
-        .isAccepted;
-  }
-
-  bool isDiplomaticAccepted(DiplomaticOrder candidate) {
-    final player = _player();
-    if (player == null) return false;
-    if (_cachedDiplomaticPrefixReplaySucceeded == false) {
-      return false;
-    }
-    final economy = _projectEconomyAfterAcceptedBuildAndWorkOrders(player);
-    final membership = _factionMembership();
-
-    if (_cachedPostDiplomaticPrefixState == null) {
-      final prefixValidator = DiplomaticOrderValidator(
-        game: game,
-        playerId: playerId,
-        initialTreasury: economy.treasury,
-        factionMembership: membership,
-      );
-      for (final existing in diplomaticOrders) {
-        final result = prefixValidator.validate(
-          existing,
-          previousRejected: false,
-        );
-        if (!result.result.isAccepted) {
-          _cachedDiplomaticPrefixReplaySucceeded = false;
-          return false;
-        }
-      }
-      _cachedDiplomaticPrefixReplaySucceeded = true;
-      _cachedPostDiplomaticPrefixState = prefixValidator
-          .capturePrefixCheckpoint();
-    }
-
-    final checkpoint = _cachedPostDiplomaticPrefixState!;
-    final candidateValidator = DiplomaticOrderValidator.fromPrefixCheckpoint(
-      game: game,
-      playerId: playerId,
-      checkpoint: checkpoint,
-      factionMembership: membership,
-    );
-    return candidateValidator
-        .validate(candidate, previousRejected: false)
-        .result
-        .isAccepted;
-  }
-
-  ({Stockpile stockpile, int treasury}) _projectEconomyAfterAcceptedBuildOrders(
-    Player player,
-  ) {
-    final cached = _cachedEconomyAfterBuildOrders;
-    if (cached != null) {
-      return cached;
-    }
-    final buildValidator = BuildOrderValidator.withProjectedEconomy(
-      game: game,
-      player: player,
-      stockpile: player.stockpile,
-      treasury:
-          player.treasury +
-          pendingRichesTreasuryDelta(
-            stockpile: player.stockpile,
-            richesCashMultiplier: game.richesCashMultiplier,
-          ),
-      workerPool: player.workerPool,
-    );
-    final builds =
-        basePrefix.buildUnitOrdersByPlayerId[playerId] ??
-        const <BuildUnitOrder>[];
-    for (final existing in builds) {
-      final result = buildValidator.validate(existing, previousRejected: false);
-      if (!result.isAccepted) {
-        final fallback = (
-          stockpile: player.stockpile,
-          treasury: player.treasury,
-        );
-        _cachedEconomyAfterBuildOrders = fallback;
-        return fallback;
-      }
-    }
-    final projected = (
-      stockpile: buildValidator.stockpile,
-      treasury: buildValidator.treasury,
-    );
-    _cachedEconomyAfterBuildOrders = projected;
-    return projected;
-  }
-
-  ({Stockpile stockpile, int treasury})
-  _projectEconomyAfterAcceptedBuildAndWorkOrders(Player player) {
-    final cached = _cachedEconomyAfterBuildAndWorkOrders;
-    if (cached != null) {
-      return cached;
-    }
-    final prefix = _ensurePostWorkPrefixState(player);
-    if (prefix == null) {
-      final afterBuild = _projectEconomyAfterAcceptedBuildOrders(player);
-      _cachedEconomyAfterBuildAndWorkOrders = afterBuild;
-      return afterBuild;
-    }
-    final projected = (stockpile: prefix.stockpile, treasury: prefix.treasury);
-    _cachedEconomyAfterBuildAndWorkOrders = projected;
-    return projected;
-  }
-
-  /// Replays accepted work orders in [basePrefix] once per validator instance,
-  /// then exposes projected economy + work-order state for candidate probes
-  /// and diplomatic projection (Refs #2394, Category B).
-  ({
-    Stockpile stockpile,
-    int treasury,
-    Set<String> seenUnitIds,
-    Set<String> devExclusive,
-  })?
-  _ensurePostWorkPrefixState(Player player) {
-    if (_cachedWorkPrefixReplaySucceeded == false) {
-      return null;
-    }
-    final cachedState = _cachedPostWorkPrefixState;
-    if (cachedState != null) {
-      return cachedState;
-    }
-    final afterBuild = _projectEconomyAfterAcceptedBuildOrders(player);
-    final works =
-        basePrefix.workOrdersByPlayerId[playerId] ?? const <WorkOrder>[];
-    final baseDev = Set<String>.from(_devExclusiveTiles());
-    if (works.isEmpty) {
-      final proj = (
-        stockpile: Stockpile(
-          quantities: Map<String, int>.from(afterBuild.stockpile.quantities),
-        ),
-        treasury: afterBuild.treasury,
-        seenUnitIds: <String>{},
-        devExclusive: baseDev,
-      );
-      _cachedWorkPrefixReplaySucceeded = true;
-      _cachedPostWorkPrefixState = proj;
-      return proj;
-    }
-    final workValidator = createWorkOrderValidator(
-      game: game,
-      player: player,
-      playerId: playerId,
-      resolution: (
-        view: view,
-        unitsById: unitsById,
-        provinceById: view.provincesById,
-      ),
-      topology: topology,
-      diplomaticOrders: diplomaticOrders,
-      tileMapByRegion: tileMapByRegion,
-      civilianDraftMoveUnitIds: _civilianDraftMoveUnitIds(),
-      devExclusiveTiles: baseDev,
-      stockpile: afterBuild.stockpile,
-      treasury: afterBuild.treasury,
-      factionMembership: _factionMembership(),
-    );
-    for (final existing in works) {
-      final result = workValidator.validate(existing, previousRejected: false);
-      if (!result.isAccepted) {
-        _cachedWorkPrefixReplaySucceeded = false;
-        return null;
-      }
-    }
-    final proj = (
-      stockpile: Stockpile(
-        quantities: Map<String, int>.from(workValidator.stockpile.quantities),
-      ),
-      treasury: workValidator.treasury,
-      seenUnitIds: {for (final w in works) w.unitId},
-      devExclusive: Set<String>.from(baseDev),
-    );
-    _cachedWorkPrefixReplaySucceeded = true;
-    _cachedPostWorkPrefixState = proj;
-    return proj;
-  }
-
-  Set<String> _civilianDraftMoveUnitIds() {
-    final cached = _cachedCivilianDraftMoveUnitIds;
-    if (cached != null) {
-      return cached;
-    }
-    final ids = <String>{};
-    final moves =
-        basePrefix.moveOrdersByPlayerId[playerId] ?? const <MoveOrder>[];
-    for (final move in moves) {
-      final unit = unitsById[move.unitId];
-      if (unit != null && unit.tileKey != null && unit.tileKey!.isNotEmpty) {
-        ids.add(move.unitId);
-      }
-    }
-    _cachedCivilianDraftMoveUnitIds = ids;
-    return ids;
-  }
-
-  Set<String> _devExclusiveTiles() {
-    final cached = _cachedDevExclusiveTiles;
-    if (cached != null) {
-      return cached;
-    }
-    final computed = devExclusiveTilesFromWorld(game.worldState, playerId);
-    _cachedDevExclusiveTiles = computed;
-    return computed;
-  }
-
-  Player? _player() => game.playerById(playerId);
 }
